This commits add the dynamic pulse detection and more.
1) The impuls protocol has a very short pulse length. Having a static
pulse length made it impossible to receive impuls codes. Other
protocols also differed in their pulse length. Making the pulse
length detection modular also made the receiving a lot more accurate.
It also reduces false positives.

Please be aware that the arctech_old protocol is broken due to a lack
of information prior to this commit.

2) Raw codes sent from the sender will now be parsed within the daemon
as if it was a received code. This enables pilight to update know
devices in the config file, but also enables users to simulate
received codes which is helpfull for protocol development.

3) A small bug was fixed in the options library which gave a
segmentation fault when a long argument was used.

4) pilight now supports the Alecto WSD-17 weather station.

5) Errors from make clean will now be suppressed.

+++ libs/hardwares/gpio.c
@@ -71,7 +71,7 @@ unsigned short gpioHwDeinit(void) {
 	return 0;
 }
 
-unsigned short gpioSend(int *code) {
+int gpioSend(int *code) {
 	unsigned short i = 0;
 
 	piHiPri(55);
